requirements.txt
google-auth==2.23.0
google-cloud-bigquery==3.11.4
jupyter==1.0.0
pandas==2.1.0
plotly==5.17.0
db-dtypes==1.1.1
requests~=2.31.0